fix: moving parsers to their own files
