+++ plugins/dependency-packages-backend-module-gem-provider/src/builder/GemComponentEntityBuilder.ts
