+++ plugins/dependency-packages-backend-module-gem-provider/src/parser/index.ts
@@ -1,0 +1,2 @@
+export { parseGemfile } from "./parseGemfile";
+export { parseGemfileLock } from "./parseGemfileLock";
