+++ plugins/dependency-packages-backend-module-gem-provider/src/retriever/GemfileLockRetriever.ts
@@ -2,8 +2,8 @@ import { LoggerService } from '@backstage/backend-plugin-api';
 import { ComponentEntity, stringifyEntityRef } from '@backstage/catalog-model';
 import { DependencyTypeRetriever } from '@voodoogq/plugin-dependency-packages-common';
 import { PACKAGE_DEPS_GEM_ANNOTATION } from '../constants';
-import { GemfileLockInfo, GemDependencyReference } from '../types';
-import fs from 'fs/promises';
+import { parseGemfile, parseGemfileLock } from '../parser';
+import { GemDependencyReference } from '../types';
 
 export interface GemfileLockRetrieverOptions {
   logger: LoggerService;
@@ -14,105 +14,6 @@ export class GemfileLockRetriever implements DependencyTypeRetriever {
 
   constructor(options: GemfileLockRetrieverOptions) {
     this.logger = options.logger;
-  }
-
-  private async parseGemfileLock(filePath: string): Promise<GemfileLockInfo> {
-    const contents = await fs.readFile(filePath.replace('file:', ''), 'utf-8');
-    const lines = contents.split('\n');
-
-    const specs: Record<string, string> = {};
-    const dependencies: string[] = [];
-    const platforms: string[] = [];
-    let bundledWith: string | null = null;
-
-    let inSpecs = false;
-    let inDependencies = false;
-    let inPlatforms = false;
-
-    for (const line of lines) {
-      const trimmed = line.trim();
-
-      if (trimmed === 'specs:') {
-        inSpecs = true;
-        inDependencies = false;
-        inPlatforms = false;
-        continue;
-      }
-      if (trimmed === 'DEPENDENCIES') {
-        inSpecs = false;
-        inDependencies = true;
-        inPlatforms = false;
-        continue;
-      }
-      if (trimmed === 'PLATFORMS') {
-        inSpecs = false;
-        inDependencies = false;
-        inPlatforms = true;
-        continue;
-      }
-      if (trimmed === 'BUNDLED WITH') {
-        inSpecs = false;
-        inDependencies = false;
-        inPlatforms = false;
-        continue;
-      }
-
-      // Actual gem specs are 4 spaces indent, NOT 6 spaces
-      if (inSpecs && line.startsWith('    ') && !line.startsWith('      ')) {
-        const match = trimmed.match(/^([a-zA-Z0-9_\-]+) \((.+?)\)/);
-        if (match) {
-          specs[match[1]] = match[2];
-        }
-      }
-
-      if (inDependencies && line.startsWith('  ')) {
-        const match = trimmed.match(/^([a-zA-Z0-9_\-]+)(?: \((.+?)\))?/);
-        if (match) {
-          const gemName = match[1];
-          const requestedVersion = match[2] ?? '';
-          dependencies.push(`${gemName}|${requestedVersion}`);
-        }
-      }
-
-      if (inPlatforms && line.startsWith('  ')) {
-        platforms.push(trimmed);
-      }
-
-      if (!inSpecs && !inDependencies && !inPlatforms && /^[ ]{3,}/.test(line) && /^\d/.test(trimmed)) {
-        // This captures the version line after "BUNDLED WITH"
-        bundledWith = trimmed;
-      }
-    }
-
-    return { specs, dependencies, platforms, bundledWith };
-  }
-
-  private async parseGemfile(filePath: string): Promise<Record<string, string[]>> {
-    const contents = await fs.readFile(filePath.replace('file:', ''), 'utf-8');
-    const lines = contents.split('\n');
-
-    const gemGroups: Record<string, string[]> = {};
-    let currentGroups: string[] = [];
-
-    for (const line of lines) {
-      const groupMatch = line.trim().match(/^group\s+(.*)?\s+do/);
-      if (groupMatch) {
-        // Extract symbols like :development, :test
-        currentGroups = groupMatch[1]
-          .split(',')
-          .map(g => g.trim().replace(/^:/, ''));
-      } else if (line.trim().startsWith('end')) {
-        currentGroups = [];
-      } else {
-        const gemMatch = line.trim().match(/^gem\s+['"]([^'"]+)['"]/);
-        if (gemMatch) {
-          const gemName = gemMatch[1];
-          gemGroups[gemName] = [...new Set([...(gemGroups[gemName] || []), ...currentGroups])];
-        }
-      }
-    }
-
-    return gemGroups;
   }
 
   public async retrieve(entities: ComponentEntity[]): Promise<GemDependencyReference[]> {
@@ -129,11 +30,9 @@ export class GemfileLockRetriever implements DependencyTypeRetriever {
       if (gemfileLocation.startsWith('file:')) {
         const gemfileLockLocation = gemfileLocation + '.lock';
         const [{ specs, dependencies, platforms, bundledWith }, gemGroups] = await Promise.all([
-          this.parseGemfileLock(gemfileLockLocation),
-          this.parseGemfile(gemfileLocation),
+          parseGemfileLock(gemfileLockLocation),
+          parseGemfile(gemfileLocation),
         ]);
-
-        console.log(specs);
 
         const gems = dependencies.map(d => {
           const [gem, requestedVersion = ''] = d.split('|');
